add makefile

+++ F1.h
@@ -3,14 +3,14 @@
 #include "PoweredVehicle.h"
 
 class F1 : public PoweredVehicle{
-	private:
+private:
 	string driver;
 	string constructor;
 	short number;
 	string engineManufacturer;
 	string trim;
-	public:
-	F1(string d = "Seb", string c = "Ferrari", string e = "FingV12", int n = 5, string t = "Quali");
+public:
+	F1(string d = "Seb", string c = "Ferrari", string e = "FingV12", short n = 5, string t = "Quali");
 	double mileageEstimate(double time);
 };
